requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.37.0
azure-cosmos==4.14.3
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
ffmpeg-python==0.2.0
filelock==3.25.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.59.2
fsspec==2026.2.0
future==1.0.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-speech==2.37.0
google-genai==1.62.0
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageHash==4.3.2
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
msrest==0.7.1
networkx==3.6.1
numpy==2.3.3
oauthlib==3.3.1
openai==2.14.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
priority==2.0.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.2.3
pytesseract==0.3.13
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyWavelets==1.9.0
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
sightengine==1.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sympy==1.14.0
tenacity==9.1.4
thefuzz==0.22.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.3.2
youtube-transcript-api==1.2.4
yt-dlp==2026.3.3
nltk
spacy
opencv-python
deepgram-sdk
aiohttp
